Use command_argument_count in mdx_main
I was having trouble compiling using macports gfortran-mp-9 due to
being unable to link to symbol iargc. I think [1] implies this is a
gfortran 77 extension, so maybe it isn't available in the mac library?
It looks like there is a replacement [2] which is standard fortran as
of '03, and using this function fixes this bug for me.

[1] https://gcc.gnu.org/onlinedocs/gfortran/IARGC.html
[2] https://gcc.gnu.org/onlinedocs/gfortran/COMMAND_005fARGUMENT_005fCOUNT.html