Remove internal usage of ISCE_HOME env variable
ISCE_HOME was only used to get the location of the default logging config.
Lots of scripts were using boilerplate to set up this config, so I added
an `isce.logging` helper module which is the same as builtin python logging
but already has the configuration defaults set up for isce.

ISCE_HOME setup is retained in the toplevel `__init__.py`
but can now be removed without affecting functionality.