Update rtcApp.py
I recommend to remove the arguments args=(self.geocode_list, self.geocode_bbox) since it passes the arguments on the runGeocode.py of the RtcProc which prevents geocoding of the sigma nought product generated from the rtcApp.py. This could be added later if someone would incorporate the geocode_list and geocode_bbox on the runGeocode.py later.