Update CMakeLists.txt
I made some changes to the CMakeLists.txt since I can't process ALOS-2 data due to the absence of alos2_slc on the db folder. Later, I've also found out that all of the folders inside the db, with the exception to the alos folder, are not included during installation.