Proposed changes to the fetchOrbit.py
After running the currently modified fetchOrbits.py, I've found out that the downloaded precise orbits were based on the date that they were processed instead of the date that they can be used. I've also modified the code to prevent downloading of restituted orbit if the precise orbit is available.

Pls. check if the changes I've made applies the orbit files to the Sentinel-1 SLC images properly.