Proposed fix on a minor bug in create_cmds.py
I'm proposing a fix to create_cmds.py since I found a minor bug in line 165 which doesn't remove the excluded dates during image stacking. Instead line 165 removes dates based on the input from the dates to be included parameter on the alosStack.xml.

I've tested the changes and it worked properly.